requirements.txt
Django==2.2.4
django-filter==2.4.0
django-widget-tweaks==1.4.8
pytz==2021.1
sqlparse==0.4.1
